d/tests: Update list_formats for extra output column in ffmpeg 7.0
The ``ffmpeg -formats`` command got an extra output column.

``ffmpeg -formats`` output header for ffmpeg 6.1:
```
File formats:
 D. = Demuxing supported
 .E = Muxing supported
```

``ffmpeg -formats`` output header for ffmpeg 7.0:
```
Formats:
 D.. = Demuxing supported
 .E. = Muxing supported
 ..d = Is a device
```

This caused problems with the ``sed`` statement and caused the list of
muxers to be empty, which caused the autopkgtests to fail.

Fix the ``sed`` statement in the ``list_formats`` function so that the
list of muxers is no longer empty.